refactor(child-process): rename module back to "child-process"
